aled last
aled last

+++ app/src/main/java/com/example/applisondage/model/ProduitModel.kt
@@ -1,6 +1,6 @@
 package com.example.applisondage.model
 
 class ProduitModel (
-    val prodId : Long = 0,
+    val prodId : Int = 0,
     val nomProd : String = ""
 )
